refactor: Update resource separation
기존 상위리소스에 하위리소스를 포함시킨 형태의 앤드포인트를 새롭게 디자인합니다. 각 리소스는 분리되어 독자적인 경로를 가지도록 수정중입니다.
