+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nest_boilerplate",
+  "name": "linkedout",
   "version": "0.0.1",
   "description": "",
   "author": "",
