+++ src/modules/follow/follow.service.ts
@@ -1,16 +1,38 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  forwardRef,
+  HttpException,
+  HttpStatus,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { FollowRepository } from './follow.repository';
-import { User } from '../../entities/user.entity';
 import { UtilsService } from '../utils/utils.service';
+import { UserSummaryResDto } from '../user/dto/response/userSummaryRes.dto';
+import { UserService } from '../user/user.service';
 
 @Injectable()
 export class FollowService {
   constructor(
     private readonly followRepository: FollowRepository,
     private readonly utilsService: UtilsService,
+    @Inject(forwardRef(() => UserService))
+    private readonly userService: UserService,
   ) {}
 
-  async follow(follower: User, following: User) {
+  async follow(followerId: number, followingId: number) {
+    if (followerId === followingId) {
+      throw new HttpException('You cannot follow yourself', HttpStatus.CONFLICT);
+    }
+
+    const followerRelation = await this.findFollowerRelation(followerId, followingId);
+    if (followerRelation) {
+      throw new HttpException('You are already following', HttpStatus.CONFLICT);
+    }
+
+    const follower = await this.userService.fetchUserEntityById(followerId);
+    const following = await this.userService.fetchUserEntityById(followingId);
+
     await this.followRepository.follow(follower, following);
   }
 
@@ -28,7 +50,12 @@ export class FollowService {
 
   async getFollowings(userId: number, page: number, limit: number) {
     const { followings, total } = await this.followRepository.findFollowings(userId, page, limit);
-    return { followings: followings, total };
+    const totalPage: number = Math.ceil(total / limit);
+
+    const followingsDto = followings.map((follow) => {
+      return this.utilsService.transformToDto(UserSummaryResDto, follow.following);
+    });
+    return { followings: followingsDto, total, totalPage, page };
   }
 
   async getAllFollowings(userId: number) {
