+++ src/modules/user/user.controller.ts
@@ -7,13 +7,12 @@ import {
   ParseIntPipe,
   Post,
   Put,
-  Query,
   Req,
   UploadedFile,
   UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
-import { ApiBody, ApiOperation, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
+import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { AuthGuard } from '@nestjs/passport';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { Request as ExpressRequest } from 'express';
@@ -23,10 +22,6 @@ import { UpdateUserReqDto } from './dto/request/updateUserReq.dto';
 import { UserResDto } from './dto/response/userRes.dto';
 import { ProfileImageUrlResDto } from './dto/response/profileImageUrlRes.dto';
 import { UserInfoSchemaDto } from './dto/schema/userInfoSchema.dto';
-import { BadgesSchemaDto } from '../badge/dto/schema/badgesSchema.dto';
-import { BadgesWithTagsSchemaDto } from '../badge/dto/schema/badgesWithTagsSchema.dto';
-import { PagingParseIntPipe } from '../../common/pipes/pagingParseInt.pipe';
-import { UserSummaryResSchemaDto } from './dto/schema/userSummaryResSchema.dto';
 import { UserSummaryWithCountSchemaDto } from './dto/schema/userSummaryWithCountSchema.dto';
 import { DeactivateReqDto } from './dto/request/deacvivateReq.dto';
 
@@ -188,155 +183,6 @@ export class UserController {
   @ApiBody({ type: UpdateUserReqDto })
   async updateUser(@Req() req: ExpressRequest, @Body() data: UpdateUserReqDto) {
     return this.userService.updateUser(req.user.id, data);
-  }
-
-  @Get('follows')
-  @ApiOperation({
-    summary: '팔로우 리스트',
-    description: `
-  현재 사용자가 팔로우하고 있는 사용자 목록을 조회합니다.
-  
-  **쿼리 파라미터:**
-  - \`page\` (number, optional): 조회할 페이지를 지정합니다. 기본값은 1입니다.
-  - \`limit\` (number, optional): 조회할 에세이 수를 지정합니다. 기본값은 20입니다.
-
-  **동작 과정:**
-  1. 사용자 ID를 기반으로 팔로우하고 있는 사용자 목록을 조회합니다.
-  2. 각 팔로우 목록을 DTO로 변환하여 반환합니다.
-
-  **주의 사항:**
-  - 요청한 사용자가 팔로우하고 있는 사용자 목록을 반환합니다.
-  - 팔로우 정보는 간략한 사용자 정보로 변환되어 반환됩니다.
-  `,
-  })
-  @ApiResponse({ status: 200, type: UserSummaryResSchemaDto })
-  @ApiQuery({ name: 'page', required: false })
-  @ApiQuery({ name: 'limit', required: false })
-  async getFollowings(
-    @Req() req: ExpressRequest,
-    @Query('page', new PagingParseIntPipe(1)) page: number,
-    @Query('limit', new PagingParseIntPipe(20)) limit: number,
-  ) {
-    return this.userService.getFollowings(req.user.id, page, limit);
-  }
-
-  @Post('follows/:userId')
-  @ApiOperation({
-    summary: '팔로우',
-    description: `
-  다른 사용자를 팔로우합니다.
-
-  **경로 파라미터:**
-  - \`userId\`: 팔로우할 사용자의 고유 ID
-
-  **동작 과정:**
-  1. 요청한 사용자의 ID와 팔로우할 사용자의 ID를 받아 팔로우 요청을 처리합니다.
-  2. 팔로우할 사용자가 존재하는지 확인합니다.
-  3. 사용자가 자신을 팔로우하려는 경우 예외를 던집니다.
-  4. 이미 팔로우하고 있는 경우 예외를 던집니다.
-  5. 팔로우 요청을 데이터베이스에 저장합니다.
-
-  **주의 사항:**
-  - 팔로우할 사용자의 ID가 유효해야 합니다.
-  - 사용자가 자신을 팔로우할 수 없습니다.
-  - 이미 팔로우 중인 경우 예외가 발생합니다.
-  `,
-  })
-  @ApiResponse({ status: 201 })
-  async follow(@Req() req: ExpressRequest, @Param('userId', ParseIntPipe) userId: number) {
-    return this.userService.follow(req.user.id, userId);
-  }
-
-  @Delete('follows/:userId')
-  @ApiOperation({
-    summary: '팔로우 취소',
-    description: `
-  다른 사용자에 대한 팔로우를 취소합니다.
-
-  **경로 파라미터:**
-  - \`userId\`: 팔로우를 취소할 사용자의 고유 ID
-
-  **동작 과정:**
-  1. 요청한 사용자의 ID와 팔로우를 취소할 사용자의 ID를 받아 팔로우 취소 요청을 처리합니다.
-  2. 팔로우 관계가 존재하는지 확인합니다.
-  3. 팔로우 관계가 존재하지 않으면 예외를 던집니다.
-  4. 팔로우 관계를 데이터베이스에서 삭제합니다.
-
-  **주의 사항:**
-  - 팔로우를 취소할 사용자의 ID가 유효해야 합니다.
-  - 팔로우 관계가 존재하지 않으면 예외가 발생합니다.
-  `,
-  })
-  @ApiResponse({ status: 204 })
-  async upFollow(@Req() req: ExpressRequest, @Param('userId', ParseIntPipe) userId: number) {
-    return this.userService.unFollow(req.user.id, userId);
-  }
-
-  @Post('badges/level/:badgeId')
-  @ApiOperation({
-    summary: '뱃지 레벨업',
-    description: `
-  사용자가 소유한 특정 뱃지의 레벨을 올립니다. 뱃지를 레벨업하기 위해서는 해당 뱃지에 필요한 경험치가 충분해야 합니다.
-
-  **경로 파라미터:**
-  - \`badgeId\`: 레벨업할 뱃지의 고유 ID
-
-  **동작 과정:**
-  1. 사용자의 ID와 레벨업할 뱃지의 ID를 받아 뱃지를 조회합니다.
-  2. 해당 뱃지가 사용자가 소유한 뱃지인지 확인합니다.
-  3. 뱃지의 경험치가 레벨업에 충분한지 확인합니다.
-  4. 경험치가 충분하지 않으면 예외를 던집니다.
-  5. 뱃지의 레벨을 올리고 경험치를 차감합니다.
-  6. 업데이트된 뱃지 정보를 데이터베이스에 저장합니다.
-
-  **주의 사항:**
-  - 뱃지를 레벨업하기 위해서는 최소한 10의 경험치가 필요합니다.
-  - 사용자가 소유하지 않은 뱃지를 레벨업하려고 하면 예외가 발생합니다.
-  `,
-  })
-  @ApiResponse({ status: 201 })
-  async levelUpBadge(@Req() req: ExpressRequest, @Param('badgeId', ParseIntPipe) badgeId: number) {
-    return this.userService.levelUpBadge(req.user.id, badgeId);
-  }
-
-  @Get('badges')
-  @ApiOperation({
-    summary: '획득한 뱃지 리스트',
-    description: `
-  사용자가 획득한 모든 뱃지 목록을 조회합니다.
-
-  **동작 과정:**
-  1. 사용자의 ID를 기반으로 해당 사용자가 획득한 모든 뱃지를 조회합니다.
-  2. 조회된 뱃지 목록을 반환합니다.
-
-  **주의 사항:**
-  - 사용자가 아직 획득하지 않은 뱃지도 기본 정보와 함께 반환됩니다.
-  `,
-  })
-  @ApiResponse({ status: 200, type: BadgesSchemaDto })
-  async userBadges(@Req() req: ExpressRequest) {
-    return this.userService.getBadges(req.user.id);
-  }
-
-  @Get('badges/detail')
-  @ApiOperation({
-    summary: '획득한 뱃지 상세 리스트',
-    description: `
-  사용자가 획득한 모든 뱃지와 해당 뱃지와 연관된 태그 목록을 조회합니다.
-
-  **동작 과정:**
-  1. 사용자의 ID를 기반으로 해당 사용자가 획득한 모든 뱃지와 연관된 태그를 조회합니다.
-  2. 조회된 뱃지와 태그 목록을 반환합니다.
-  3. 사용자가 아직 획득하지 않은 뱃지도 기본 정보와 함께 반환됩니다.
-
-  **주의 사항:**
-  - 사용자가 아직 획득하지 않은 뱃지도 기본 정보와 함께 반환됩니다.
-  - 각 뱃지에는 연관된 태그 목록이 포함됩니다.
-  `,
-  })
-  @ApiResponse({ status: 200, type: BadgesWithTagsSchemaDto })
-  async userBadgesWithTags(@Req() req: ExpressRequest) {
-    return this.userService.getBadgeWithTags(req.user.id);
   }
 
   @Get('summary')
