+++ src/modules/user/user.module.ts
@@ -1,3 +1,4 @@
+import * as strategies from '../../common/guards/strategies';
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { JwtModule } from '@nestjs/jwt';
@@ -6,14 +7,11 @@ import { MailModule } from '../mail/mail.module';
 import { AwsModule } from '../aws/aws.module';
 import { EssayModule } from '../essay/essay.module';
 import { UtilsModule } from '../utils/utils.module';
-import { FollowModule } from '../follow/follow.module';
-import { BadgeModule } from '../badge/badge.module';
 import { UserController } from './user.controller';
 import { UserRepository } from './user.repository';
 import { UserService } from './user.service';
 import { User } from '../../entities/user.entity';
 import { Essay } from '../../entities/essay.entity';
-import * as strategies from '../../common/guards/strategies';
 import { NicknameModule } from '../nickname/nickname.module';
 import { DeactivationReason } from '../../entities/deactivationReason.entity';
 import { BullModule } from '@nestjs/bull';
@@ -37,11 +35,9 @@ import { UserProcessor } from './user.processor';
       inject: [ConfigService],
     }),
     AuthModule,
-    FollowModule,
     MailModule,
     AwsModule,
     UtilsModule,
-    BadgeModule,
     NicknameModule,
     forwardRef(() => EssayModule),
   ],
